Revert "Code Refactoring"
This reverts commit bfec31cc8ee750cc010aa2b4110f92f8c47fce71.
