+++ audit-event-generator/src/main/resources/log4j.properties
@@ -9,6 +9,3 @@ log4j.appender.file.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss,SSS} %-5p %-
 
 # Suppress the irrelevant (wrong) warnings from the Netty channel handler
 log4j.logger.org.apache.flink.shaded.akka.org.jboss.netty.channel.DefaultChannelPipeline=ERROR, file
-
-
-
