+++ audit-event-generator/src/main/scala/org/sunbird/job/auditevent/domain/Event.scala
@@ -1,4 +1,4 @@
-package org.sunbird.job.transaction.domain
+package org.sunbird.job.auditevent.domain
 
 import org.apache.commons.lang3.StringUtils
 import org.sunbird.job.domain.reader.JobRequest
