+++ audit-event-generator/src/main/scala/org/sunbird/job/auditevent/task/AuditEventGeneratorConfig.scala
@@ -1,4 +1,4 @@
-package org.sunbird.job.transaction.task
+package org.sunbird.job.auditevent.task
 
 import java.util
 import com.typesafe.config.Config
