+++ pom.xml
@@ -8,12 +8,8 @@
 	<prerequisites>
 		<maven>3.0.0</maven>
 	</prerequisites>
-    <modules>
-        <module>transaction-event-processor</module>
-        <module>newMod</module>
-    </modules>
 
-    <groupId>org.sunbird</groupId>
+	<groupId>org.sunbird</groupId>
 	<artifactId>knowledge-platform-jobs</artifactId>
 	<version>1.0</version>
 	<packaging>pom</packaging>
