--- transaction-event-processor/src/main/resources/log4j.properties
@@ -1,14 +1,0 @@
-# log4j.appender.file=org.apache.log4j.FileAppender
-log4j.appender.file=org.apache.log4j.RollingFileAppender
-log4j.appender.file.file=audit-event-generator.log
-log4j.appender.file.append=true
-log4j.appender.file.layout=org.apache.log4j.PatternLayout
-log4j.appender.file.MaxFileSize=256KB
-log4j.appender.file.MaxBackupIndex=4
-log4j.appender.file.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss,SSS} %-5p %-60c %x - %m%n
-
-# Suppress the irrelevant (wrong) warnings from the Netty channel handler
-log4j.logger.org.apache.flink.shaded.akka.org.jboss.netty.channel.DefaultChannelPipeline=ERROR, file
-
-
-
