--- transaction-event-processor/src/main/resources/transaction-event-processor.conf
@@ -1,25 +1,0 @@
-include "base-config.conf"
-
-kafka {
-  input.topic = "sunbirddev.learning.graph.events"
-  output.topic = "sunbirddev.telemetry.raw"
-  groupId = "sunbirddev-transaction-event-processor-group"
-}
-
-task {
-  consumer.parallelism = 1
-  parallelism = 1
-  producer.parallelism = 1
-  window.time = 60
-}
-
-schema {
-  basePath = "https://sunbirddev.blob.core.windows.net/sunbird-content-dev/schemas/local"
-}
-
-channel.default = "org.sunbird"
-
-job {
-    audit-event-generator = true
-    audit-history-indexer = true
-}
