--- transaction-event-processor/src/test/resources/logback-test.xml
@@ -1,16 +1,0 @@
-<configuration debug="true">
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <root level="debug">
-        <appender-ref ref="STDOUT" />
-    </root>
-
-    <logger name="org.cassandraunit" level="ERROR"/>
-    <logger name="com.datastax.driver.core" level="ERROR"/>
-    <logger name="com.outworkers.phantom" level="ERROR"/>
-    <logger name="org.apache.cassandra" level="off"/>
-</configuration>
